# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Service",
        "description": "API by for longnguyen",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/branch": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get List Branch",
                "description": "Display a listing of the resource.",
                "operationId": "getBranches",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "An error occurred while Display a listing the branch."
                    }
                }
            },
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Create Branch",
                "description": "Store a newly created resource in storage.",
                "operationId": "createBranch",
                "requestBody": {
                    "description": "Create Branch",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_name",
                                    "branch_phone",
                                    "branch_email",
                                    "branch_address",
                                    "branch_status"
                                ],
                                "properties": {
                                    "branch_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Branch 1"
                                    },
                                    "branch_phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "0123456789"
                                    },
                                    "branch_email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "example@example.com"
                                    },
                                    "branch_address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Address 1"
                                    },
                                    "branch_status": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred while creating the Branch."
                    }
                }
            }
        },
        "/api/v1/branch/{id}": {
            "delete": {
                "tags": [
                    "Branch"
                ],
                "summary": "Delete Branch",
                "description": "Delete Branch",
                "operationId": "deleteBranch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of branch to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch deleted successfully"
                    },
                    "500": {
                        "description": "An error occurred while deleting the Branch."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Branch"
                ],
                "summary": "Update Branch",
                "description": "Update Branch",
                "operationId": "updateBranch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of branch to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Branch",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_name",
                                    "branch_phone",
                                    "branch_email",
                                    "branch_address",
                                    "branch_status"
                                ],
                                "properties": {
                                    "branch_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Branch 1 updated"
                                    },
                                    "branch_phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "01234567899"
                                    },
                                    "branch_email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "exampleupdate@example.com"
                                    },
                                    "branch_address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Address 1 updated"
                                    },
                                    "branch_status": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occurred while updating the Branch."
                    }
                }
            }
        },
...
    }
}